requirements.txt
# requirements.txt
pandas
numpy
torch
torchvision
streamlit
geopy
requests
python-dotenv
sqlalchemy
psycopg2-binary
# Geospatial libraries
rasterio
geopandas